Added installation of VS code

+++ apt_install.sh
@@ -22,12 +22,36 @@ install htop
 install tree
 install dropbox
 install octave
+install software-properties-common 
+install apt-transport-https 
+install wget
 
 install gnome-tweaks
 install chrome-gnome-shell
 install gnome-shell-extensions
 
 # Install chrome:
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-sudo apt install ./google-chrome-stable_current_amd64.deb
-rm ./google-chrome-stable_current_amd64.deb
+which google-chrome &> /dev/null
+
+if [ $? -ne 0 ]; then
+  echo "Installing: google-chrome..."
+  wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+  sudo apt install ./google-chrome-stable_current_amd64.deb
+  rm ./google-chrome-stable_current_amd64.deb    
+else
+  echo "Already installed: google-chrome"
+fi
+
+
+# Install VS code:
+which code &> /dev/null
+
+if [ $? -ne 0 ]; then
+  echo "Installing: VS code..."
+  wget -q https://packages.microsoft.com/keys/microsoft.asc -O- | sudo apt-key add -
+  sudo add-apt-repository "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main"
+  sudo apt update
+  sudo apt install code  
+else
+  echo "Already installed: VS code"
+fi
